Enum, write n bytes, gyro fs set function added
